[x86] 32-bit Bootstrap code in Xvisor
This will only work with Grub2 as it only it can load
ELF64 compatible binary in protected mode. Purpose of
this patch is to put intial code for bootstrap before
64-bit code can be executed.

It right now only checks if long mode (64-bit) is supported
on the processor or not. Setting up page tables for
long mode and far jumping to 64-bit is still pending.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>